+++ app/src/main/java/org/juanro/autumandu/gui/util/NewRefuelingSnackbar.java
@@ -110,7 +110,7 @@ public class NewRefuelingSnackbar {
 
             return fuelConsumption.computeFuelConsumption(
                     volumeSinceFullRefueling,
-                    thisMileage - mileageOfLastFullRefueling);
+                    (float) thisMileage - mileageOfLastFullRefueling);
         } else {
             return -1;
         }
